ISS4827: Lifted processComponents and several other utility functions into their own file hierarchies
