+++ packages/mtw-wml/ts/standardize/index.ts
@@ -1,35 +1,27 @@
-import { SchemaTag, isSchemaConditionStatement, isSchemaCondition, isSchemaConditionFallthrough, isImportable, SchemaWithKey, isSchemaImport, isSchemaCharacter, isSchemaRoom, isSchemaFeature, isSchemaKnowledge, isSchemaBookmark, isSchemaMap, isSchemaMessage, isSchemaMoment, isSchemaTheme, isSchemaVariable, isSchemaComputed, isSchemaAction, isSchemaImage, isSchemaAsset, isSchemaMeta, SchemaAssetTag, isSchemaExport, isSchemaRemove, isSchemaWithKey, isSchemaReplace, isSchemaReplaceMatch, isSchemaReplacePayload, isSchemaExit, isSchemaLink } from "../schema/baseClasses"
+import { SchemaTag, isSchemaConditionStatement, isSchemaCondition, isSchemaConditionFallthrough, isImportable, SchemaWithKey, isSchemaImport, isSchemaAsset, isSchemaMeta, SchemaAssetTag, isSchemaExport, isSchemaRemove, isSchemaWithKey, isSchemaExit, isSchemaLink } from "../schema/baseClasses"
 import { GenericTree, GenericTreeNode, GenericTreeNodeFiltered, treeNodeTypeguard } from "../tree/baseClasses"
-import { isStandardAction, isStandardBookmark, isStandardCharacter, isStandardComputed, isStandardFeature, isStandardImage, isStandardKnowledge, isStandardMap, isStandardMessage, isStandardMoment, isStandardNDJSON, isStandardRemove, isStandardReplace, isStandardRoom, isStandardTheme, isStandardVariable, MergeConflictError, SerializeNDJSONMixin, StandardFormSubsetRequest, StandardFormSubsetRequestExit, StandardFormSubsetRequestFull, standardFormSubsetRequestPriority, StandardNDJSON } from "./baseClasses"
+import { isStandardNDJSON, MergeConflictError, SerializeNDJSONMixin, StandardFormSubsetRequest, StandardFormSubsetRequestExit, StandardFormSubsetRequestFull, standardFormSubsetRequestPriority, StandardNDJSON } from "./baseClasses"
 import { excludeUndefined } from "../lib/lists"
-import { isStandardComponent, isStandardForm, StandardComponentData, StandardComponentNonEditData, StandardFormData, StandardRemoveData, StandardReplaceData } from "./components/dataTypes"
+import { isStandardComponent, isStandardForm, StandardComponentData, StandardComponentNonEditData, StandardFormData } from "./components/dataTypes"
 import { unique } from "../list"
 import SchemaTagTree from "../tagTree/schema"
-import { TagListItem, TagTreeMatchOperation } from "../tagTree"
 import applyEdits from "../schema/treeManipulation/applyEdits"
 import StandardRoom, { StandardRoomPayload } from "./components/room"
 import StandardFeature, { StandardFeaturePayload } from "./components/feature"
 import StandardKnowledge, { StandardKnowledgePayload } from "./components/knowledge"
-import StandardBookmark from "./components/bookmark"
 import StandardMap from "./components/map"
-import StandardMessage from "./components/message"
-import StandardMoment from "./components/moment"
-import StandardTheme from "./components/theme"
-import StandardVariable from "./components/variable"
-import StandardComputed from "./components/computed"
-import StandardAction from "./components/action"
-import StandardImage from "./components/image"
-import StandardCharacter from "./components/character"
 import { isSchemaTreeNode } from "./components/utils"
-import { unwrapSubject, wrappedNodeTypeGuard } from "../schema/utils"
+import { wrappedNodeTypeGuard } from "../schema/utils"
 import { HasDescription, HasName, HasShortName } from "./components/abstract"
-import { isLegalKey, nodeFromWML } from "./utils"
+import { isLegalKey, nodeFromWML, removeNDJSONOnlyProperties } from "./utils"
 import { StandardBaseData } from "./components/dataTypes/abstract"
-import { StandardComponentExport, StandardComponentImport } from "./components/dataTypes/metaData"
 import { StandardComponent } from "./components/component"
-import { KeyPayload } from "./components/key"
-import { deepEqual, objectFilterEntries, objectMap } from "../lib/objects"
-import { ExportItemContent, ExportItemRemove, ExportItemReplace, ImportItemContent, ImportItemRemove, ImportItemReplace, StandardExportItem, StandardImportItem } from "./components/metaData"
+import { deepEqual, objectMap } from "../lib/objects"
+import { ExportItemContent, ExportItemRemove, ExportItemReplace, ImportItemContent, ImportItemRemove, ImportItemReplace } from "./components/metaData"
+import processComponents, { ComponentProcessingTemplate } from "./processComponents"
+import { StandardRemove, StandardReplace } from "./edits"
+import { standardComponentFactory } from "./componentFactory"
+import importExportFromTree from "./importExportFromTree"
 
 export const assertTypeguard = <T extends any, G extends T>(value: T, typeguard: (value) => value is G): G => {
     if (typeguard(value)) {
@@ -108,528 +100,6 @@ export const standardComponentSortOrder = (componentA: StandardComponent, compon
     }
 }
 
-//
-// standardNonEditComponentFactory takes an incoming argument that can apply to one of the non-edit StandardComponent classes,
-// finds the correct constructor, and creates the sub-typed class
-//
-export const standardNonEditComponentFactory = (arg: StandardComponentData | GenericTreeNode<SchemaTag>): StandardComponent | undefined => {
-
-    if ((!isSchemaTreeNode(arg) && isStandardCharacter(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaCharacter)(arg))) {
-        return new StandardCharacter(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardRoom(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaRoom)(arg))) {
-        return new StandardRoom(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardFeature(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaFeature)(arg))) {
-        return new StandardFeature(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardKnowledge(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaKnowledge)(arg))) {
-        return new StandardKnowledge(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardBookmark(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaBookmark)(arg))) {
-        return new StandardBookmark(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardMap(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaMap)(arg))) {
-        return new StandardMap(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardMessage(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaMessage)(arg))) {
-        return new StandardMessage(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardMoment(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaMoment)(arg))) {
-        return new StandardMoment(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardTheme(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaTheme)(arg))) {
-        return new StandardTheme(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardVariable(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaVariable)(arg))) {
-        return new StandardVariable(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardComputed(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaComputed)(arg))) {
-        return new StandardComputed(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardAction(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaAction)(arg))) {
-        return new StandardAction(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardImage(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaImage)(arg))) {
-        return new StandardImage(arg)
-    }
-    return undefined
-}
-
-const removeNDJSONOnlyProperties = (props: StandardComponentData & SerializeNDJSONMixin): Omit<StandardComponentData & SerializeNDJSONMixin, 'universalKey' | 'from' | 'exportAs'> => {
-    return Object.assign({}, 
-        ...Object.entries(props)
-            .filter(([key]) => (!['universalKey', 'from', 'exportAs'].includes(key)))
-            .map(([key, value]) => ({ [key]: value }))
-    )
-}
-
-//
-// StandardRemove class provides a class that contains a matching StandardComponent to be removed. Note that merge
-// methods at this level do NOT contain the functionality to handle component-level edits ... that is included
-// at the StandardForm level, rather than on the individual component classes.
-//
-export class StandardRemove implements StandardComponent {
-    _key: KeyPayload;
-    _match: StandardComponent;
-    tag: SchemaWithKey["tag"] | 'Remove' | 'Replace' = 'Remove' as const;
-    constructor(props: string | StandardRemoveData | GenericTreeNode<SchemaTag> | StandardRemove) {
-        if (props instanceof StandardRemove) {
-            this._key = props._key
-            this._match = props._match.clone()
-            return
-        }
-        if (isSchemaTreeNode(props) || typeof props === 'string') {
-            const node = typeof props === 'string'
-                ? nodeFromWML(props)
-                : props
-            if (!treeNodeTypeguard(isSchemaRemove)(node)) {
-                throw new Error(`Schema mismatch in StandardRemove constructor call.`)
-            }
-            const child = node.children[0]
-            if (!treeNodeTypeguard(isSchemaWithKey)(child)) {
-                throw new Error(`No key found in StandardRemove constructor call.`)
-            }
-            this._key = new KeyPayload(child.data.key)
-            const match = standardNonEditComponentFactory(child)
-            if (!match) {
-                throw new Error('No payload found in StandardRemove constructor call.')
-            }
-            this._match = match
-            this._key._universalKey = match.universalKey
-            return
-        }
-        const match = standardNonEditComponentFactory(props.component)
-        if (!match) {
-            throw new Error('No payload found in StandardRemove constructor call.')
-        }
-        this._match = match
-        this._key = new KeyPayload({ key: match.key, universalKey: match.universalKey })
-    }
-
-    get key() { return this._key.key }
-    get universalKey() { return this._key.universalKey }
-    get fileName() { return this._key.fileName }
-    get import() { return this._match.import }
-    get export() { return this._match.export }
-
-    referencedKeys(): { key: string; referenceType: "Link" | "Position" | "Exit" | "Direct" | "Dependency" }[] {
-        return this._match.referencedKeys()
-    }
-
-    clone(): StandardRemove {
-        return new StandardRemove(this)
-    }
-
-    mapContents(callback): StandardRemove {
-        const returnValue = new StandardRemove(this)
-        returnValue._match = returnValue._match.mapContents(callback)
-        return returnValue
-    }
-
-    toJSON(): StandardRemoveData {
-        return {
-            key: this.key,
-            tag: 'Remove',
-            component: this._match.toJSON() as StandardComponentNonEditData & SerializeNDJSONMixin
-        }
-    }
-
-    toNDJSON(): StandardComponentData & SerializeNDJSONMixin {
-        return this.toJSON()
-    }
-
-    get schema(): GenericTreeNode<SchemaTag> {
-        return {
-            data: { tag: 'Remove' },
-            children: [this._match.schema]
-        }
-    }
-
-    merge(incoming: StandardComponent): StandardComponent | undefined {
-        throw new Error('StandardRemove types cannot be directly merged')
-    }
-
-    withKey(key: string): StandardComponent {
-        const returnValue = new StandardRemove(this.schema)
-        returnValue._match = this._match.withKey(key)
-        returnValue._key._key = returnValue._match.key
-        returnValue._key._fileName = returnValue._match.fileName
-        returnValue._key._universalKey = returnValue._match.universalKey
-        return returnValue
-    }
-
-    withUniversalKey(key: string | undefined): StandardComponent {
-        const returnValue = new StandardRemove(this.schema)
-        returnValue._match = this._match.withUniversalKey(key)
-        returnValue._key._key = returnValue._match.key
-        returnValue._key._fileName = returnValue._match.fileName
-        returnValue._key._universalKey = returnValue._match.universalKey
-        return returnValue
-    }
-
-    withFileName(key: string | undefined): StandardComponent {
-        const returnValue = new StandardRemove(this.schema)
-        returnValue._match = this._match.withFileName(key)
-        returnValue._key._key = returnValue._match.key
-        returnValue._key._fileName = key
-        returnValue._key._universalKey = returnValue._match.universalKey
-        return returnValue
-    }
-
-    withImport(importData: StandardImportItem | StandardComponentImport | undefined): StandardComponent {
-        const returnValue = new StandardRemove(this.schema)
-        returnValue._match = this._match.withImport(importData)
-        returnValue._key._key = returnValue._match.key
-        returnValue._key._fileName = returnValue._match.fileName
-        returnValue._key._universalKey = returnValue._match.universalKey
-        return returnValue
-    }
-
-    withExport(exportData: StandardExportItem | StandardComponentExport | string | undefined): StandardComponent {
-        const returnValue = new StandardRemove(this.schema)
-        returnValue._match = this._match.withExport(exportData)
-        returnValue._key._key = returnValue._match.key
-        returnValue._key._fileName = returnValue._match.fileName
-        returnValue._key._universalKey = returnValue._match.universalKey
-        return returnValue
-    }
-}
-
-//
-// StandardReplace class provides a class that contains a matching StandardComponent to be removed. Note that merge
-// methods at this level do NOT contain the functionality to handle component-level edits ... that is included
-// at the StandardForm level, rather than on the individual component classes.
-//
-export class StandardReplace implements StandardComponent {
-    _key: KeyPayload;
-    _match: StandardComponent;
-    _payload: StandardComponent
-    tag: SchemaWithKey["tag"] | 'Remove' | 'Replace' = 'Replace' as const;
-    constructor(props: string | StandardReplaceData | GenericTreeNode<SchemaTag> | StandardReplace) {
-        if (props instanceof StandardReplace) {
-            this._key = props._key
-            this._match = props._match.clone()
-            this._payload = props._payload.clone()
-            return
-        }
-        if (isSchemaTreeNode(props) || typeof props === 'string') {
-            const node = typeof props === 'string'
-                ? nodeFromWML(props)
-                : props
-            if (!treeNodeTypeguard(isSchemaReplace)(node)) {
-                throw new Error(`Schema mismatch in StandardReplace constructor call.`)
-            }
-            const matchNode = node.children.find(treeNodeTypeguard(isSchemaReplaceMatch))?.children?.[0]
-            const match = matchNode ? standardNonEditComponentFactory(matchNode) : undefined
-            const payloadNode = node.children.find(treeNodeTypeguard(isSchemaReplacePayload))?.children?.[0]
-            const payload = payloadNode ? standardNonEditComponentFactory(payloadNode) : undefined
-            if (!match) {
-                throw new Error('No match found in StandardReplace constructor call.')
-            }
-            if (!payload) {
-                throw new Error('No payload found in StandardReplace constructor call.')
-            }
-            if (!(match.key === payload.key && match.tag === payload.tag)) {
-                throw new Error('Match and payload mistmatch in StandardReplace constructor call.')
-            }
-            this._match = match
-            this._payload = payload
-            this._key = new KeyPayload({ key: match.key, universalKey: match.universalKey })
-            return
-        }
-        const match = standardNonEditComponentFactory(props.match)
-        if (!match) {
-            throw new Error('No payload found in StandardRemove constructor call.')
-        }
-        const payload = standardNonEditComponentFactory(props.payload)
-        if (!payload) {
-            throw new Error('No payload found in StandardRemove constructor call.')
-        }
-        if (!(match.key === payload.key && match.tag === payload.tag)) {
-            throw new Error('Match and payload mistmatch in StandardReplace constructor call.')
-        }
-        this._match = match
-        this._payload = payload
-        this._key = new KeyPayload({ key: match.key, universalKey: match.universalKey })
-    }
-
-    get key() { return this._key.key }
-    get universalKey() { return this._key.universalKey }
-    get fileName() { return this._key.fileName }
-    get import() { return this._match.import }
-    get export() { return this._match.export }
-
-    clone(): StandardReplace {
-        return new StandardReplace(this)
-    }
-
-    mapContents(callback): StandardReplace {
-        const returnValue = new StandardReplace(this)
-        returnValue._match = returnValue._match.mapContents(callback)
-        returnValue._payload = returnValue._payload.mapContents(callback)
-        return returnValue
-    }
-
-    toJSON(): StandardReplaceData {
-        return {
-            key: this.key,
-            tag: 'Replace',
-            match: this._match.toJSON() as StandardComponentNonEditData & SerializeNDJSONMixin,
-            payload: this._payload.toJSON() as StandardComponentNonEditData & SerializeNDJSONMixin
-        }
-    }
-
-    toNDJSON(): StandardComponentData & SerializeNDJSONMixin {
-        return this.toJSON()
-    }
-
-    get schema(): GenericTreeNode<SchemaTag> {
-        return {
-            data: { tag: 'Replace' },
-            children: [
-                { data: { tag: 'ReplaceMatch' }, children: [this._match.schema] },
-                { data: { tag: 'ReplacePayload' }, children: [this._payload.schema] }
-            ]
-        }
-    }
-
-    merge(incoming: StandardComponent): StandardComponent | undefined {
-        if (!(incoming instanceof StandardReplace)) {
-            throw new Error('Type mismatch in StandardReplace merge')
-        }
-        if (!(deepEqual(removeNDJSONOnlyProperties(this._payload.toJSON()), removeNDJSONOnlyProperties(incoming._match.toJSON())))) {
-            throw new MergeConflictError()
-        }
-        return new StandardReplace({
-            key: this.key,
-            tag: 'Replace',
-            match: this._match.toJSON() as StandardComponentNonEditData,
-            payload: incoming._payload.toJSON() as StandardComponentNonEditData
-        }).withUniversalKey(this.universalKey)
-    }
-
-    referencedKeys(): { key: string; referenceType: "Link" | "Position" | "Exit" | "Direct" | "Dependency" }[] {
-        return [
-            ...this._match.referencedKeys(),
-            ...this._payload.referencedKeys()
-        ]
-    }
-
-    withKey(key: string): StandardComponent {
-        const returnValue = new StandardReplace(this.schema)
-        returnValue._match = this._match.withKey(key)
-        returnValue._payload = this._match.withKey(key)
-        returnValue._key._universalKey = key
-        returnValue._key._fileName = this.fileName
-        return returnValue
-    }
-
-    withUniversalKey(key: string | undefined): StandardComponent {
-        const returnValue = new StandardReplace(this.schema)
-        returnValue._match = this._match.withUniversalKey(key)
-        returnValue._payload = this._match.withUniversalKey(key)
-        returnValue._key._universalKey = key
-        returnValue._key._fileName = this.fileName
-        return returnValue
-    }
-
-    withFileName(key: string | undefined): StandardComponent {
-        const returnValue = new StandardReplace(this.schema)
-        returnValue._match = this._match.withFileName(key)
-        returnValue._payload = this._payload.withFileName(key)
-        returnValue._key._fileName = key
-        returnValue._key._universalKey = this._key._universalKey
-        return returnValue
-    }
-
-    withImport(importData: StandardImportItem | StandardComponentImport | undefined): StandardComponent {
-        const returnValue = new StandardReplace(this.schema)
-        returnValue._match = this._match.withImport(importData)
-        returnValue._payload = this._payload.withImport(importData)
-        returnValue._key._key = returnValue._match.key
-        returnValue._key._fileName = returnValue._match.fileName
-        returnValue._key._universalKey = returnValue._match.universalKey
-        return returnValue
-    }
-
-    withExport(exportData: StandardExportItem | StandardComponentExport | string | undefined): StandardComponent {
-        const returnValue = new StandardReplace(this.schema)
-        returnValue._match = this._match.withExport(exportData)
-        returnValue._payload = this._payload.withExport(exportData)
-        returnValue._key._key = returnValue._match.key
-        returnValue._key._fileName = returnValue._match.fileName
-        returnValue._key._universalKey = returnValue._match.universalKey
-        return returnValue
-    }
-
-}
-
-//
-// standardComponentFactory takes an incoming argument that can apply to any of the StandardComponent classes (including Remove and Replace),
-// finds the correct constructor, and creates the sub-typed class
-//
-export const standardComponentFactory = (arg: StandardComponentData | GenericTreeNode<SchemaTag>): StandardComponent | undefined => {
-    if ((!isSchemaTreeNode(arg) && isStandardRemove(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaRemove)(arg))) {
-        return new StandardRemove(arg)
-    }
-    if ((!isSchemaTreeNode(arg) && isStandardReplace(arg)) || (isSchemaTreeNode(arg) && treeNodeTypeguard(isSchemaReplace)(arg))) {
-        return new StandardReplace(arg)
-    }
-    return standardNonEditComponentFactory(arg)
-}
-
-//
-// Utility function to create exportItemById and importItemById objects, then use them to inform the
-// creation of this._byId
-//
-const importExportFromTree = (tree: GenericTree<SchemaTag>): { importItemById: Record<string, StandardImportItem>; exportItemById: Record<string, StandardExportItem> } => {
-    const mergeExportIntoEntries = (previous: Record<string, StandardExportItem>, incoming: Record<string, StandardExportItem>): Record<string, StandardExportItem> => {
-        const [[key, incomingItem]] = Object.entries(incoming)
-        const baseItem = previous[key]
-        const mergedItem = baseItem ? baseItem.merge(incomingItem) : incomingItem
-        if (mergedItem) {
-            return {
-                ...previous,
-                [key]: mergedItem
-            }
-        }
-        else {
-            return objectFilterEntries(previous, ([checkKey]) => (checkKey !== key))
-        }
-    }
-    const exportItemById = tree.filter(wrappedNodeTypeGuard(isSchemaExport))
-        .reduce<Record<string, StandardExportItem>>((previous, node) => {
-            if (treeNodeTypeguard(isSchemaRemove)(node)) {
-                const child = node.children[0]
-                if (child && treeNodeTypeguard(isSchemaExport)(child)) {
-                    return child.children
-                        .map(unwrapSubject)
-                        .filter(excludeUndefined)
-                        .map(({ data }) => (data))
-                        .filter(isImportable)
-                        .map(({ key, as }) => ({ [key]: new ExportItemRemove(as ?? key) }))
-                        .reduce<Record<string, StandardExportItem>>(mergeExportIntoEntries, previous)
-                }
-            }
-            if (treeNodeTypeguard(isSchemaReplace)(node)) {
-                throw new Error('Top-level replace of Export tags not yet implemented')
-            }
-            if (treeNodeTypeguard(isSchemaExport)(node)) {
-                const removeActions = node.children
-                    .filter(treeNodeTypeguard(isSchemaRemove))
-                    .map(unwrapSubject)
-                    .filter(excludeUndefined)
-                    .map(({ data }) => (data))
-                    .filter(isImportable)
-                    .map(({ key, as }) => ({ [key]: new ExportItemRemove(as ?? key) }))
-                const replaceActions = node.children
-                    .filter(treeNodeTypeguard(isSchemaReplace))
-                    .map(({ children }) => ({
-                        match: children.find(treeNodeTypeguard(isSchemaReplaceMatch))?.children?.[0],
-                        payload: children.find(treeNodeTypeguard(isSchemaReplacePayload))?.children?.[0]
-                    }))
-                    .map(({ match, payload }) => (
-                        (match && treeNodeTypeguard(isImportable)(match) && payload && treeNodeTypeguard(isImportable)(payload))
-                            ? [{ [match.data.key]: new ExportItemReplace(match.data.as ?? match.data.key, payload.data.as ?? payload.data.key) }]
-                            : []
-                    ))
-                    .flat(1)
-                const contentActions = node.children
-                    .filter(treeNodeTypeguard(isImportable))
-                    .map(({ data }) => (data))
-                    .map(({ key, as }) => ({ [key]: new ExportItemContent(as ?? key) }))
-                return contentActions.reduce(
-                    mergeExportIntoEntries,
-                    replaceActions.reduce(
-                        mergeExportIntoEntries,
-                        removeActions.reduce(
-                            mergeExportIntoEntries,
-                            previous
-                        )
-                    )
-                )
-            }
-            return previous
-        }, {})
-    const mergeImportIntoEntries = (previous: Record<string, StandardImportItem>, incoming: Record<string, StandardImportItem>): Record<string, StandardImportItem> => {
-        const [[key, incomingItem]] = Object.entries(incoming)
-        const baseItem = previous[key]
-        const mergedItem = baseItem ? baseItem.merge(incomingItem) : incomingItem
-        if (mergedItem) {
-            return {
-                ...previous,
-                [key]: mergedItem
-            }
-        }
-        else {
-            return objectFilterEntries(previous, ([checkKey]) => (checkKey !== key))
-        }
-    }
-    const importItemById = tree.filter(wrappedNodeTypeGuard(isSchemaImport))
-        .reduce<Record<string, StandardImportItem>>((previous, node) => {
-            if (treeNodeTypeguard(isSchemaRemove)(node)) {
-                const child = node.children[0]
-                if (child && treeNodeTypeguard(isSchemaImport)(child)) {
-                    return child.children
-                        .map(unwrapSubject)
-                        .filter(excludeUndefined)
-                        .map(({ data }) => (data))
-                        .filter(isImportable)
-                        .map(({ key, from }) => ({ [key]: new ImportItemRemove(child.data.from, from ?? key) }))
-                        .reduce<Record<string, StandardImportItem>>(mergeImportIntoEntries, previous)
-                }
-            }
-            if (treeNodeTypeguard(isSchemaReplace)(node)) {
-                throw new Error('Top-level replace of Export tags not yet implemented')
-            }
-            if (treeNodeTypeguard(isSchemaImport)(node)) {
-                const removeActions: Record<string, StandardImportItem>[] = node.children
-                    .filter(treeNodeTypeguard(isSchemaRemove))
-                    .map(unwrapSubject)
-                    .filter(excludeUndefined)
-                    .map(({ data }) => (data))
-                    .filter(isImportable)
-                    .map(({ key, from }) => ({ [key]: new ImportItemRemove(node.data.from, from ?? key) }))
-                const replaceActions = node.children
-                    .filter(treeNodeTypeguard(isSchemaReplace))
-                    .map(({ children }) => ({
-                        match: children.find(treeNodeTypeguard(isSchemaReplaceMatch))?.children?.[0],
-                        payload: children.find(treeNodeTypeguard(isSchemaReplacePayload))?.children?.[0]
-                    }))
-                    .map(({ match, payload }) => (
-                        (match && treeNodeTypeguard(isImportable)(match) && payload && treeNodeTypeguard(isImportable)(payload))
-                            ? [{ [match.data.key]: new ImportItemReplace(
-                                { assetId: node.data.from, fromKey: match.data.from ?? match.data.key },
-                                { assetId: node.data.from, fromKey: payload.data.from ?? payload.data.key }
-                            ) }]
-                            : []
-                    ))
-                    .flat(1)
-                const contentActions = node.children
-                    .filter(treeNodeTypeguard(isImportable))
-                    .map(({ data }) => (data))
-                    .map(({ key, from }) => ({ [key]: new ImportItemContent(node.data.from, from ?? key) }))
-                return contentActions.reduce(
-                    mergeImportIntoEntries,
-                    replaceActions.reduce(
-                        mergeImportIntoEntries,
-                        removeActions.reduce(
-                            mergeImportIntoEntries,
-                            previous
-                        )
-                    )
-                )
-            }
-            return previous
-        }, {})
-    return { importItemById, exportItemById }
-}
-
-type ComponentProcessingTemplate = {
-    key: SchemaWithKey["tag"];
-}
-
 export class StandardForm {
     _key: string;
     _byId: Record<string, StandardComponent>;
@@ -690,155 +160,6 @@ export class StandardForm {
                 ? nodeFromWML(args)
                 : args
 
-            const keysByComponentTypeFactory = (tagTree: SchemaTagTree) => (tag: SchemaWithKey["tag"]) => {
-                const keysExtract = (imported: boolean) => (
-                    tagTree
-                        .filter({ and: [{ match: tag }, imported ? { match: 'Import' } : { not: { match: 'Import' } }] })
-                        .prune({ after: { match: tag } })
-                        .prune({ before: { match: tag } })
-                        .tree
-                        .map(({ data }) => {
-                            if (data.tag !== tag) {
-                                throw new Error('standardizeSchema tag mismatch')
-                            }
-                            if (imported && isImportable(data)) {
-                                return data.as ?? data.key
-                            }
-                            return data.key
-                        })
-                )
-                return unique(keysExtract(true), keysExtract(false)).sort()
-            }
-            const { importItemById, exportItemById } = importExportFromTree(node.children)
-            const standardizeComponentTagType = (componentTemplates: ComponentProcessingTemplate[], tagTree: SchemaTagTree): void => {
-                //
-                // Loop through each tag in standard order
-                //
-                const anyKeyedComponent: TagTreeMatchOperation<SchemaTag> = { or: componentTemplates.map(({ key }) => ({ match: key })) }
-                componentTemplates.forEach((processingTemplate) => {
-                    const { key: tag } = processingTemplate
-                    //
-                    // Loop through each key present for that tag
-                    //
-                    const keys = keysByComponentTypeFactory(tagTree)(tag)
-                    keys.forEach((key) => {
-                        //
-                        // Aggregate and reorder all top-level information
-                        //
-                        const nodeMatch: TagTreeMatchOperation<SchemaTag> = { match: ({ data }, stack) => (data.tag === tag && ('as' in data ? data.as === key : data.key === key)) }
-                        const nodeMatchImport: TagTreeMatchOperation<SchemaTag> = { match: ({ data }, stack) => (data.tag === tag && (((Boolean(stack.find(isSchemaImport)) && isImportable(data)) ? data.as ?? data.key : data.key) === key)) }
-                        const editTag: TagTreeMatchOperation<SchemaTag> = { or: [{ match: 'Replace' }, { match: 'Remove' }] }
-                        const adjustTagTree = (tagTree: SchemaTagTree, nodeMatch: TagTreeMatchOperation<SchemaTag>): SchemaTagTree => {
-                            const prunedTagTree = tagTree
-                                .prune({ after: { sequence: [nodeMatch, anyKeyedComponent] } })
-                                .reorderFunctional(
-                                    [{ match: tag }, { match: 'Replace'}, { match: 'ReplaceMatch' }, { match: 'ReplacePayload' }, { match: 'Remove' }, { match: 'Name' }, { match: 'ShortName' }, { match: 'Description' }, { match: 'Summary' }, { match: 'If' }, { match: 'Statement' }, { match: 'Fallthrough' }],
-                                    (tagItem) => {
-                                        const isEditTag = (value: TagListItem<SchemaTag, {}>): boolean => (['Replace', 'ReplaceMatch', 'ReplacePayload', 'Remove'].includes(value.data.tag))
-                                        const isConditionalTag = (value: TagListItem<SchemaTag, {}>): boolean => (['If', 'Statement', 'Fallthrough'].includes(value.data.tag))
-                                        const { componentTags, valueTags, conditionalTags } = tagItem.reduce<{ componentTags: TagListItem<SchemaTag>[]; valueTags: TagListItem<SchemaTag>[]; conditionalTags: TagListItem<SchemaTag>[]; matchedAlready: boolean }>((previous, subItem) => {
-                                            if (subItem.data.tag === tag) {
-                                                return {
-                                                    ...previous,
-                                                    componentTags: [...previous.componentTags, subItem],
-                                                    matchedAlready: true
-                                                }
-                                            }
-                                            if (isEditTag(subItem)) {
-                                                if (previous.matchedAlready) {
-                                                    return {
-                                                        ...previous,
-                                                        valueTags: [...previous.valueTags, subItem]
-                                                    }
-                                                }
-                                                else {
-                                                    return {
-                                                        ...previous,
-                                                        componentTags: [...previous.componentTags, subItem]
-                                                    }
-                                                }
-                                            }
-                                            if (isConditionalTag(subItem)) {
-                                                return {
-                                                    ...previous,
-                                                    conditionalTags: [...previous.conditionalTags, subItem]
-                                                }
-                                            }
-                                            else {
-                                                return {
-                                                    ...previous,
-                                                    valueTags: [...previous.valueTags, subItem]
-                                                }
-                                            }
-                                        }, { componentTags: [], valueTags: [], conditionalTags: [], matchedAlready: false })
-                                        const relativeOrder: Partial<Record<SchemaTag["tag"], number>> = {
-                                            Remove: 1,
-                                            Replace: 1,
-                                            ReplaceMatch: 2,
-                                            ReplacePayload: 2,
-                                            [tag]: 3,
-                                            Name: 4,
-                                            ShortName: 4,
-                                            Description: 4,
-                                            Summary: 4
-                                        }
-                                        const sortInPlace = (tags: TagListItem<SchemaTag>[]): TagListItem<SchemaTag>[] => (
-                                            [...tags].sort((a, b) => ((relativeOrder[a.data.tag] ?? Infinity) - (relativeOrder[b.data.tag] ?? Infinity)))
-                                        )
-                                        return [...sortInPlace(componentTags), ...sortInPlace(valueTags), ...conditionalTags]
-                                    }
-                                )
-                                .filter({ not: { sequence: [{ or: [ { match: 'Remove' }, { match: 'Replace' }] }, { or: [{ match: 'Import' }, { match: 'Export' }] }] }})
-                                .prune({ and: [{ before: nodeMatch }, { not: { or: [editTag, { after: editTag }] }}] })
-                                .prune({ or: [{ match: 'Import' }, { match: 'Export' }] })
-                            switch(tag) {
-                                case 'Room':
-                                    return prunedTagTree.prune({ or: [{ match: 'Map' }, { match: 'Position' }]})
-                                case 'Map':
-                                    return tagTree
-                                        .prune({ or: [{ and: [{ after: { sequence: [nodeMatch, anyKeyedComponent] } }, { not: { match: 'Position'} }] }, { match: 'Import' }, { match: 'Export' }] })
-                                        .reordered([{ match: tag }, { or: [{ match: 'Name' }, { match: 'Description' }] }, { or: [{ match: 'Room' }, { connected: [{ match: 'If' }, { or: [{ match: 'Statement' }, { match: 'Fallthrough' }]}] } ]}])
-                                        .filter({ or: [{ and: [{ match: 'Room' }, { or: [{ match: 'Position' }, { match: 'Exit' }]}]}, { not: { match: 'Room' }}]})
-                                        .prune({ before: nodeMatch })
-                            }
-                            return prunedTagTree
-                        }
-
-                        const filteredTagTree = adjustTagTree(tagTree.filter({ and: [nodeMatch, { not: { match: 'Import' } }] }), nodeMatch)
-                        const importedTagTree = adjustTagTree(tagTree.filter({ and: [nodeMatchImport, { match: 'Import' }] }), nodeMatchImport)
-
-                        const adjustedTree = [
-                            ...(applyEdits(adjustTagTree(importedTagTree, nodeMatch).tree)
-                                .map((item) => (
-                                    (treeNodeTypeguard(isImportable)(item) && item.data.as)
-                                        ? { ...item, data: { ...item.data, key: item.data.as } }
-                                        : item
-                                ))),
-                            ...applyEdits(adjustTagTree(filteredTagTree, nodeMatch).tree)
-                        ]
-                        adjustedTree.forEach((item) => {
-                            const standardItem = standardComponentFactory(item)
-                            if (standardItem) {
-                                const base = this._byId[standardItem.key]
-                                if (base) {
-                                    const merged = base.merge(standardItem as any)
-                                    if (merged) {
-                                        const mergedImport = base.import && standardItem.import ? base.import.merge(standardItem.import) : base.import ?? standardItem.import
-                                        const mergedExport = base.export && standardItem.export ? base.export.merge(standardItem.export) : base.export ?? standardItem.export
-                                        this._byId[standardItem.key] = merged.withImport(mergedImport).withExport(mergedExport)
-                                    }
-                                    else {
-                                        delete this._byId[standardItem.key]
-                                    }
-                                }
-                                else {
-                                    this._byId[standardItem.key] = standardItem.withImport(importItemById[standardItem.key]).withExport(exportItemById[standardItem.key])
-                                }
-                            }
-                        })
-                    })
-                })
-            }
             this._byId = {}
             this._metaData = []
 
@@ -875,7 +196,14 @@ export class StandardForm {
                     { key: 'Action' }
                 ]
         
-                standardizeComponentTagType(componentTemplates, tagTree)
+                const { importItemById, exportItemById } = importExportFromTree(asset.children)
+                processComponents({
+                    componentTemplates,
+                    tagTree,
+                    byId: this._byId,
+                    importItemById,
+                    exportItemById
+                })
                 return
             }
         }
